Fixed flex bug in menu. In loadable components, moved common functions and members to abstract class.
