+++ src/app/header/header.component.scss
@@ -121,7 +121,6 @@
 
     .navbar-brand {
       float: none;
-      display: inline-flex;
       vertical-align: middle;
       padding: 0;
